Adding BlobStoreStatsTest
