+++ ambry-store/src/main/java/com.github.ambry.store/BlobStore.java
@@ -16,7 +16,6 @@ package com.github.ambry.store;
 import com.codahale.metrics.Timer;
 import com.github.ambry.config.StoreConfig;
 import com.github.ambry.utils.FileLock;
-import com.github.ambry.utils.Pair;
 import com.github.ambry.utils.Time;
 import java.io.File;
 import java.io.IOException;
@@ -26,7 +25,6 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Set;
-import java.util.SortedMap;
 import java.util.concurrent.ScheduledExecutorService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
